Added extract_exons_gtf.py script
Nexons has an option to create gtf output files, but these do not separate out the exons on to individual rows. 
This script should take a gtf output file from nexons and create a new gtf file with each exon on a separate line.
This should be compatible with IGV. Information from the first and last splice sites currently get thrown away.